refactor: crud material
